requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Survival analysis
lifelines>=0.27.0
scikit-survival>=0.17.0
pycox>=0.2.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0
tensorflow>=2.7.0
keras>=2.7.0

# Causal inference
econml>=0.13.0
causalml>=0.13.0
dowhy>=0.8.0

# Reinforcement learning
stable-baselines3>=1.6.0
gym>=0.21.0

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Data processing
pyarrow>=6.0.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
dill>=0.3.0
cloudpickle>=2.0.0

# Configuration
pyyaml>=6.0
configparser>=5.2.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Model interpretation
shap>=0.40.0
lime>=0.2.0
eli5>=0.11.0

# Hyperparameter optimization
optuna>=2.10.0
hyperopt>=0.2.0
scikit-optimize>=0.9.0

# Time series
tslearn>=0.5.0
sktime>=0.11.0

# Clinical validation
krippendorff>=0.5.0
fleiss-kappa>=1.0.0

# Performance monitoring
mlflow>=1.20.0
wandb>=0.12.0

# Data validation
great-expectations>=0.15.0
pandera>=0.11.0

